requirements.txt
alembic==1.0.10
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.1
boto3==1.9.188
botocore==1.12.188
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
dataset==1.1.2
decorator==4.4.2
defusedxml==0.6.0
docutils==0.14
ecdsa==0.15
entrypoints==0.3
envs==1.3
idna==2.8
importlib-metadata==1.3.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.4
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
Mako==1.0.10
MarkupSafe==1.1.1
mistune==0.8.4
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
packaging==19.2
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
psycopg2-binary==2.8.2
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.8
pycparser==2.20
Pygments==2.6.1
PyMySQL==0.9.3
pyparsing==2.4.5
pyrsistent==0.15.7
pytest==5.3.1
python-box==3.4.0
python-dateutil==2.8.0
python-dotenv==0.10.5
python-editor==1.0.4
python-jose==3.1.0
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
requests==2.22.0
rsa==4.0
s3transfer==0.2.1
Send2Trash==1.5.0
six==1.12.0
SQLAlchemy==1.3.3
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.3
watchtower==0.7.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==0.6.0
